adatta a macchina personale

+++ src/utils.py
@@ -24,8 +24,9 @@ def plot_time_lines(ax, zz):
 
 def directory_cleanup(region):
     """Remove old plots and create directory for new ones."""
-    dir_path = f"/Users/giorgioclauser/Documents/covid_plts/{region}"
-    shutil.rmtree(dir_path)
+    dir_path = f"/Users/GiorgioClauser 1/Documents/covid_plt/{region}"
+    if os.path.isdir(dir_path):
+        shutil.rmtree(dir_path)
     os.mkdir(dir_path)
     return dir_path
